# server/openapi.yaml
openapi: 3.0.3
info:
  title: 蔓蔓点菜 API
  description: |
    蔓蔓点菜微信小程序后端 REST API。

    **认证方式**：所有标注 requireAuth 的端点需在请求头传递 `X-WX-OPENID: <用户openid>`。
    管理端点需额外传递 `X-WX-ADMIN: true`。

    **通用错误 envelope**：
    ```json
    { "error": { "code": "ERROR_CODE", "message": "中文描述" }, "requestId": "uuid" }
    ```
  version: 1.0.0
  contact:
    name: 蔓蔓点菜开发组
servers:
  - url: https://express-xxx.sh.run.tcloudbase.com
    description: 云托管生产环境（私有链路，仅 wx.cloud.callContainer 可访问）
  - url: http://127.0.0.1:8080
    description: 本地开发
paths:
...
  /api/v1/meal-plans:
    get:
      operationId: listMealPlans
      summary: 查询当前用户的点菜记录
      security:
        - AuthHeader: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: 起始日期（包含），格式 yyyy-MM-dd
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: 结束日期（包含），格式 yyyy-MM-dd
      responses:
        '200':
          description: 点菜记录列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MealPlan'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createMealPlan
      summary: 首次提交点菜计划
      description: |
        幂等支持：若请求头 `Idempotency-Key` 与历史请求相同，但请求体不同，
        返回 `409 IDEMPOTENCY_CONFLICT`；相同 key + 相同 body 返回原结果（不重复创建）。
      security:
        - AuthHeader: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
          description: 幂等键，相同 key + 相同 body 重放返回原结果
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMealPlanPayload'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MealPlan'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/meal-plans/{id}:
    put:
      operationId: updateMealPlan
      summary: 修改已存在的点菜计划
      description: |
        **乐观锁**：必须携带当前 `version` 字段；旧 version 返回 `409 VERSION_CONFLICT`。
        日期和餐次不可修改。
      security:
        - AuthHeader: []
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMealPlanPayload'
      responses:
        '200':
          description: 更新后的点菜计划
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MealPlan'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'

  # -------------------------------------------------------------------------
  # 通知订阅与配额
  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    AuthHeader:
      type: apiKey
      in: header
      name: X-WX-OPENID
      description: 微信用户 OpenID，由小程序 cloud.getWXContext() 获取并透传
    AdminHeader:
      type: apiKey
      in: header
      name: X-WX-OPENID
      description: 管理员 OpenID（需在 ADMIN_OPENIDS 环境变量中配置）
    InternalToken:
      type: apiKey
      in: header
      name: X-Notify-Token
      description: notify-admin 与 Express 服务共享的访问令牌（NOTIFY_API_TOKEN 环境变量）
  parameters:
...
    MealPlanId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 点菜计划 ID
...
  schemas:
...
    MealPlanItem:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: 菜品 ID
    MealPlan:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date
          description: 就餐日期（yyyy-MM-dd）
        mealType:
          type: string
          enum: [breakfast, lunch, dinner]
        items:
          type: array
          items:
            $ref: '#/components/schemas/MealPlanItem'
        note:
          type: string
          maxLength: 100
        version:
          type: integer
          description: 乐观锁版本号
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMealPlanPayload:
      type: object
      required: [date, mealType, items]
      properties:
        date:
          type: string
          format: date
          description: 就餐日期（yyyy-MM-dd），支持今天及未来30天
        mealType:
          type: string
          enum: [breakfast, lunch, dinner]
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/MealPlanItem'
        note:
          type: string
          maxLength: 100
    UpdateMealPlanPayload:
      type: object
      required: [items, version]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/MealPlanItem'
        note:
          type: string
          maxLength: 100
        version:
          type: integer
          description: 当前版本号（乐观锁）
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
      example:
        error:
          code: "VALIDATION_ERROR"
          message: "缺少必填字段"
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "至少选择一道菜"
    Unauthorized:
      description: 未认证（缺少 X-WX-OPENID）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "缺少用户身份"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "无权操作此资源"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "记录不存在"
    VersionConflict:
      description: 乐观锁版本冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VERSION_CONFLICT"
              message: "版本冲突，请刷新后重试"
    IdempotencyConflict:
      description: 幂等键冲突（相同 key 不同请求体）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "IDEMPOTENCY_CONFLICT"
              message: "幂等键冲突：请求体不一致"